Fix empty hook name if ArrayPattern is transpiled
This occurs when the ArrayPattern node is transpiled.

``` js
var _useState = useState(0),
  foo = _useState[0],
  setFoo = _useState[1];
```

We'll try to detect the above pattern for `useState` and `useReducer`
and pick the name from the second variable declarator on match.
If the pattern cannot be detected, we'll fall back to the hook name
itself, e.g. "useState".